- Reimplement Pharo-Launcher-Core-BenComan.70 without using an additional class.
The idea is that the previous method #reset was useless except during opening of the Pharo Launcher. So I renamed it to #initializeTemplatesAndGroups and the code always send this message during the initialization of the object (the Pharo Launcher does not have to request it anymore).

This means we have now 2 methods: #refresh (that refreshes everything) and #refreshLocal (that refreshes only the local repositories).