These changes resolve https://pharo.fogbugz.com/f/cases/12640.
However this is dependent on update to Spec from SLICE-Issue-12677. So until that is integrated into Pharo mainline, this package has been named to keep it out of PharoLauncher CI.  When 12677 is integrated, the changes in this package should moved into PharoLauncher mainline (without this package as an ancestor) and this package deleted from SmalltalkHub.