This PR add two of Peter's customizations for the launcher.
The first one is to let the user select a display strategy for the "last modified" date and to add the possibility to display it like "3 weeks ago".

The second one is to add a setting to colour the date to know which image was launch recently really easily.

Fixes https://github.com/pharo-project/pharo-launcher/issues/226