Speed up a lot the time to fetch the image names.
On my machine, with multiple images it takes 1.1sec to fetch the image names before. With this PR it goes down to 47ms. 

This participate to reduce the important of #412. 

I still think it is not enough because to add 47ms of overhead each time we type a character is already too much. And I have a SSD and in the future I could get mor images.