Speedup vm fetching.
Instead of ensuring the rights of each executables when creating the vms, I now create all the VM then ensure they are executables in one command. 

This also fix a bug. If we updated a VM and already had the executable file in cache, it did not ensured that the new executable was executable.