[Bug]: When Setting Image Location, Compare Unresolved Locations
Previously, a resolved current location (e.g. '/Users/me/myfolder' asFileReference) was compared to a possibly unresolved new location (e.g. FileLocator home / 'myfolder'), so even if the current and new location was the *same* FileLocation, a migration would be done. Resolving both wouldn't work, because changing from a hardcoded location to a dynamic one that just happens to currrently point to the same place seems to be a valid operation as well. The safest course seems to be to trust the user that they mean "set it to this exactly" and not "set it to something somehow equivalent to this"