VM and Image - User Customizability
Introduce hooks for custom VM and Image Subclasses, which can specify their own e.g.:
- Executable name
- Default image arguments (GT needs this because the image immediately closes after startup unless `'--no-quit'` is supplied)
- `pharo.version` interpretation

As a working use case example, PharoLauncher-GToolkit allows predownloaded Gtoolkit VMs and images to be used via Launcher.

NB: Didn't add PharoLauncher-GToolkit to the baseline. Should we? There are no dependencies except Launcher, so it can be easily loaded on top of Launcher. Maybe we should at least define the package so that it can be loaded via MetaC API?